+++ app.js
@@ -6,11 +6,11 @@
   var routers = {};
 
   routers.stockCheck = express.Router();
-  routers.textSearch = express.Router()
+  routers.search = express.Router();
   require('./config.js')(app, routers);
 
   require('./server/stock/stock_routes.js')(routers.stockCheck);
-  require('./server/search/search_routes.js')(routers.textSearch);
+  require('./server/search/search_routes.js')(routers.search);
 
   module.exports = exports = app;
 }());