+++ server/search/search_routes.js
@@ -3,7 +3,10 @@
   var controller = require('./search_controllers.js');
 
   module.exports = exports = function (router) {
-    router.route('')
+    router.route('/advanced')
       .get(controller.textSearch);
+
+    router.route('/simple')
+        .get(controller.search);
   };
 })();