+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - added jshint (conor.fennell)
 - fixed the path to ./server/stock/stock_routes.js (conor.fennell)
 - added text search route (brian.lough)
+- added search route for search by product number (brian.lough)